+++ src/chunks/coords.rs
@@ -6,7 +6,7 @@ use num_integer::Integer;
 pub struct ChunkCoords(pub i32, pub i32);
 
 impl ChunkCoords {
-    pub fn from_pos(pos: Vector3<f32>) -> Self {
+    pub fn from_pos(pos: Vector3<f64>) -> Self {
         let (x, _, z) = block_coords(pos);
         Self::from_block(x, z)
     }
@@ -42,6 +42,6 @@ impl ChunkCoords {
     }
 }
 
-pub fn block_coords(pos: Vector3<f32>) -> (i32, i32, i32) {
+pub fn block_coords(pos: Vector3<f64>) -> (i32, i32, i32) {
     (pos.x.floor() as i32, pos.y.floor() as i32, pos.z.floor() as i32)
 }
