+++ src/entities/components.rs
@@ -1,7 +1,7 @@
 use nalgebra::Vector3;
 
 #[derive(Clone, Copy, Debug, Default)]
-pub struct Position(pub Vector3<f32>);
+pub struct Position(pub Vector3<f64>);
 
 #[derive(Clone, Copy, Debug, Default)]
 pub struct Rotation(pub f32, pub f32);
