+++ src/net/packets/builder.rs
@@ -40,7 +40,7 @@ impl PacketBuilder {
         self
     }
 
-    pub fn add_position_delta(&mut self, delta: f32) -> &mut Self {
+    pub fn add_position_delta(&mut self, delta: f64) -> &mut Self {
         let delta = (delta * 4096.0) as i16;
         self.add_bytes(&delta.to_be_bytes());
         self
